fix: remove reference to wagmiConfig causing desync in modal initialization resulting in email connector not being setup
