+++ apps/laboratory/src/pages/library/wagmi-all.tsx
@@ -13,7 +13,7 @@ import { siweConfig } from '../../utils/SiweUtils'
 
 const queryClient = new QueryClient()
 
-export const wagmiConfig = defaultWagmiConfig({
+const wagmiConfig = defaultWagmiConfig({
   chains: WagmiConstantsUtil.chains,
   projectId: ConstantsUtil.ProjectId,
   metadata: ConstantsUtil.Metadata,
