+++ apps/laboratory/src/pages/library/wagmi.tsx
@@ -11,7 +11,7 @@ import { WagmiModalInfo } from '../../components/Wagmi/WagmiModalInfo'
 
 const queryClient = new QueryClient()
 
-export const wagmiConfig = defaultWagmiConfig({
+const wagmiConfig = defaultWagmiConfig({
   chains: WagmiConstantsUtil.chains,
   projectId: ConstantsUtil.ProjectId,
   metadata: ConstantsUtil.Metadata,
