+++ apps/laboratory/src/utils/WagmiConstants.ts
@@ -1,3 +1,4 @@
+import { defaultWagmiConfig } from '@web3modal/wagmi/react/config'
 import {
   arbitrum,
   aurora,
@@ -15,6 +16,7 @@ import {
   optimismSepolia,
   type Chain
 } from 'wagmi/chains'
+import { ConstantsUtil } from './ConstantsUtil'
 
 export const WagmiConstantsUtil = {
   chains: [
@@ -34,3 +36,14 @@ export const WagmiConstantsUtil = {
     optimismSepolia
   ] as [Chain, ...Chain[]]
 }
+
+export const wagmiConfig = defaultWagmiConfig({
+  chains: WagmiConstantsUtil.chains,
+  projectId: ConstantsUtil.ProjectId,
+  metadata: ConstantsUtil.Metadata,
+  enableEmail: true,
+  auth: {
+    socials: ['google', 'x', 'discord', 'apple', 'github']
+  },
+  ssr: true
+})
